fix: turn into hook
